pyproject.toml
[tool.poetry]
name = "beekeeper"
version = "0.1.0"
description = "Exporting SupportBee data for better integration with other ticketing tools"
authors = ["Jorge Sanz <jsanz@carto.com>"]
license = "MIT"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.dependencies]
python = "^3.6"
requests = "^2.21"
PyYAML = "^5.1"
click = "^7.0"
pathos = "^0.2.3"

[tool.poetry.dev-dependencies]
jupyterlab = "^0.35.4"
ipdb = "^0.12.0"
ipython = "^7.4"

[tool.poetry.scripts]
beekeeper = "beekeeper.cli:cli"


